// sdk/src/api/openapi.ts
export interface components {
    schemas: {
// ...
        InteractionEventType: "survey_presented" | "question_viewed" | "answer_selected" | "answer_changed" | "answer_deselected" | "text_focused" | "text_edited" | "text_blurred" | "validation_blocked" | "question_skipped" | "question_not_applicable" | "navigated_next" | "navigated_back" | "question_left" | "survey_backgrounded" | "survey_foregrounded" | "survey_dismissed" | "survey_completed";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}